feat: schema/API hardening — alembic, integer money, backups, fixes
DB (alembic 도입):
- 금액 컬럼 Float→BigInteger(KRW 원 단위), USING round()로 데이터 변환
- transactions.date 인덱스 + stats/목록 쿼리를 extract()→date 범위 조건으로
- budgets.category unique 제약 (기존 중복은 최신 행 유지 후 제거)
- 기존 create_all DB는 시작 시 baseline(0001) 자동 stamp 후 upgrade

API:
- PATCH /transactions 부호 정규화 (지출 양수 입력 시 음수 저장)
- 카테고리 직접 수정 시 category_confirmed + category_cache 학습
- 카테고리 rename 시 거래/할부/구독/예산/캐시 문자열 참조 일괄 갱신
- GET /subscriptions 부작용 제거 → POST /subscriptions/sync 분리
- PATCH 전반 exclude_none→exclude_unset (null 되돌리기 가능)

Infra:
- docker compose에 로컬 postgres 추가 (NAS 운영 DB 접근 차단)
- k8s: DB 백업 CronJob(pg_dump 일일/30일 보관), 구독 sync CronJob,
  api TZ=Asia/Seoul, livenessProbe, .dockerignore
- PWA manifest 아이콘(icon-192/512.png) 추가
- CLAUDE.md 현행화 (Ollama→Anthropic API, alembic 절차)

+++ api/schemas.py
@@ -11,14 +11,14 @@ from models import AccountType, TransactionType
 class AccountCreate(BaseModel):
     name: str
     type: AccountType
-    balance: float = 0.0
+    balance: int = 0
     color: str = "#6366f1"
     icon: str = "💳"
 
 
 class AccountUpdate(BaseModel):
     name: Optional[str] = None
-    balance: Optional[float] = None
+    balance: Optional[int] = None
     color: Optional[str] = None
     icon: Optional[str] = None
 
@@ -27,7 +27,7 @@ class AccountResponse(BaseModel):
     id: int
     name: str
     type: AccountType
-    balance: float
+    balance: int
     color: str
     icon: str
     created_at: datetime
@@ -39,7 +39,7 @@ class AccountResponse(BaseModel):
 
 class TransactionCreate(BaseModel):
     account_id: Optional[int] = None
-    amount: float
+    amount: int
     description: str = ""
     type: TransactionType
     date: Optional[datetime] = None
@@ -50,7 +50,7 @@ class TransactionUpdate(BaseModel):
     category: Optional[str] = None
     subcategory: Optional[str] = None
     category_confirmed: Optional[bool] = None
-    amount: Optional[float] = None
+    amount: Optional[int] = None
     date: Optional[datetime] = None
     account_id: Optional[int] = None
 
@@ -60,7 +60,7 @@ class TransactionResponse(BaseModel):
     account_id: Optional[int]
     installment_id: Optional[int]
     subscription_id: Optional[int]
-    amount: float
+    amount: int
     description: str
     category: Optional[str]
     subcategory: Optional[str]
@@ -77,7 +77,7 @@ class TransactionResponse(BaseModel):
 
 class InstallmentCreate(BaseModel):
     name: str
-    total_amount: float
+    total_amount: int
     total_months: int
     annual_interest_rate: Optional[float] = None
     start_year: int
@@ -88,7 +88,7 @@ class InstallmentCreate(BaseModel):
 
 class InstallmentUpdate(BaseModel):
     name: Optional[str] = None
-    total_amount: Optional[float] = None
+    total_amount: Optional[int] = None
     total_months: Optional[int] = None
     annual_interest_rate: Optional[float] = None
     start_year: Optional[int] = None
@@ -100,14 +100,14 @@ class InstallmentUpdate(BaseModel):
 class InstallmentResponse(BaseModel):
     id: int
     name: str
-    total_amount: float
+    total_amount: int
     total_months: int
     annual_interest_rate: Optional[float]
     start_year: int
     start_month: int
     category: Optional[str]
     subcategory: Optional[str]
-    monthly_amount: float
+    monthly_amount: int
     created_at: datetime
 
     model_config = {"from_attributes": True}
@@ -142,7 +142,7 @@ class CategoryUpdate(BaseModel):
 
 class SubscriptionCreate(BaseModel):
     name: str
-    amount: float
+    amount: int
     cycle: str = "monthly"
     billing_day: int = 1
     start_year: int
@@ -153,7 +153,7 @@ class SubscriptionCreate(BaseModel):
 
 class SubscriptionUpdate(BaseModel):
     name: Optional[str] = None
-    amount: Optional[float] = None
+    amount: Optional[int] = None
     cycle: Optional[str] = None
     billing_day: Optional[int] = None
     category: Optional[str] = None
@@ -164,7 +164,7 @@ class SubscriptionUpdate(BaseModel):
 class SubscriptionResponse(BaseModel):
     id: int
     name: str
-    amount: float
+    amount: int
     cycle: str
     billing_day: int
     category: Optional[str]
@@ -180,13 +180,13 @@ class SubscriptionResponse(BaseModel):
 
 class BudgetCreate(BaseModel):
     category: str
-    monthly_amount: float
+    monthly_amount: int
 
 
 class BudgetResponse(BaseModel):
     id: int
     category: str
-    monthly_amount: float
+    monthly_amount: int
     created_at: datetime
 
     model_config = {"from_attributes": True}
@@ -233,7 +233,7 @@ class DailyStat(BaseModel):
 class TopTransaction(BaseModel):
     id: int
     description: str
-    amount: float
+    amount: int
     category: Optional[str]
     subcategory: Optional[str]
     date: datetime
